Basic Ghost Elements
Ghost is stored as a set of X, Y, and Velocity values. Every frame we
grab the values from player and store them into tables. When a certain
amount of time has passed (100 frames) we start having the ghost move
and do stuff. This is simple, but it should work how we want it to. We
should create our own class for ghosts. That or we should add a new
platform type, but that's stupid. Do a class. I'm going to bed.